refactor(itun): decompose CrawlerStorageSection into sub-components
Split CrawlerStorageSection (686 -> 166 lines) into 4 focused
sub-components: EntityStorageItem, CustomStorageItem,
CustomItemDetailModal, CustomStorageDialog. All co-located in the
games directory.

Closes #125 #126
